Last Changes to Report

+++ Bluetooth/BSN/src/bm_blemesh_model_handler.c
@@ -65,7 +65,7 @@ static void button0_cb(){
 	err = bt_mesh_onoff_cli_set_unack(&on_off_cli, NULL, &set);
 	//err = bt_mesh_onoff_cli_set(&on_off_cli, NULL, &set, NULL);
 	#ifdef BENCHMARK_CLIENT
-	bm_log_append_ram((bm_message_info) {on_off_cli.tid,synctimer_getSyncTime(),0,on_off_cli.model->pub->ttl,0,addr,on_off_cli.model->pub->addr,on_off_cli.model->pub->addr,on_off_cli.pub.msg->len});
+	bm_log_append_ram((bm_message_info) {(uint16_t)on_off_cli.tid,synctimer_getSyncTime(),(uint64_t)0,on_off_cli.model->pub->ttl,0,addr,on_off_cli.model->pub->addr,on_off_cli.model->pub->addr,on_off_cli.pub.msg->len});
 	#endif
 	/*
 	printk("Sent TID %u\n",on_off_cli.tid);
@@ -95,7 +95,7 @@ static void led_set(struct bt_mesh_onoff_srv *srv, struct bt_mesh_msg_ctx *ctx,
 	rsp->present_on_off = set->on_off;
 	// Log the Event
 	#ifdef BENCHMARK_SERVER
-	bm_log_append_ram((bm_message_info) {srv->prev_transaction.tid+1,synctimer_getSyncTime(),0,(uint8_t)BLE_MESH_TTL-ctx->recv_ttl,ctx->recv_rssi,ctx->addr,ctx->recv_dst,ctx->recv_dst,srv->model->pub->msg->len});
+	bm_log_append_ram((bm_message_info) {(uint16_t)srv->prev_transaction.tid+1,synctimer_getSyncTime(),0,(uint8_t)BLE_MESH_TTL-ctx->recv_ttl,(uint8_t)ctx->recv_rssi,ctx->addr,ctx->recv_dst,ctx->recv_dst,srv->model->pub->msg->len});
 	#endif
 	/*
 	printk("Recv TID %u\n",srv->prev_transaction.tid+1);
